refactor(platform-wallet): remove PrivateKeyData/KeyStorage and a vestigial re-export shim
Two related cleanups in the same file.

1. `PrivateKeyData` and `KeyStorage` were never constructed or read —
   `PrivateKeyData::` returns zero hits across all of packages/, including
   rs-platform-wallet-ffi. The doc claimed "the IdentityKeysChangeSet apply path
   constructs one per replay, the FFI key-preview path uses one internally";
   `apply_identity_key_entry` touches neither. `key_storage.rs` is renamed to
   `identity_status.rs` — it now holds only `IdentityStatus` and `DpnsNameInfo`.

2. `state::managed_identity` re-exported `block_time` and `key_storage` under
   the old path so "external users can still reach them" — a comment describing
   a past move, not the current layout. Ten sites in the tree (including
   rs-platform-wallet-ffi/src/memory_explorer.rs) reached the types through the
   shim instead of `crate::wallet::identity::types::*`. All retargeted, shim
   removed.

Import-path changes only — no C ABI impact.
