+++ packages/rs-platform-wallet/src/lib.rs
@@ -83,8 +83,8 @@ pub use wallet::identity::{
     derive_contact_payment_addresses, derive_contact_xpub, pubkey_binds_expected_key_data,
     unmask_account_reference, BlockTime, ContactProfileEntry, ContactRequest, ContactXpubData,
     DashPayProfile, DashPayState, DpnsNameInfo, EstablishedContact, IdentityLocation,
-    IdentityManager, IdentityStatus, KeyStorage, ManagedIdentity, PrivateKeyData, ProfileUpdate,
-    RegistrationIndex, DEFAULT_CONTACT_GAP_LIMIT,
+    IdentityManager, IdentityStatus, ManagedIdentity, ProfileUpdate, RegistrationIndex,
+    DEFAULT_CONTACT_GAP_LIMIT,
 };
 pub use wallet::masternode_withdrawal::{
     MasternodeWithdrawalKey, MasternodeWithdrawalKeys, MasternodeWithdrawalRequest,
